Move helpers to tests/__init__.py
